split scaffolder into separate file
