+++ base/index.js
@@ -1,8 +1,8 @@
+const _ = require('lodash');
 const fs = require('fs');
 const path = require('path');
 const mkdirp = require('mkdirp');
-const mustache = require('mustache');
-const _ = require('lodash');
+const scaffold = require('../lib/scaffold');
 
 module.exports = (opts) => {
   const things = [
@@ -30,19 +30,8 @@ module.exports = (opts) => {
     {p: 'tests/routes', n: 'schema'},
   ].map(thing => _.extend({p: '', d: thing.p || '', t: thing.n, e: 'js'}, thing));
 
-  things.forEach(thing => {
-    const templatePath = path.join(__dirname, thing.p, thing.n+'.mustache');
-    const template = fs.readFileSync(templatePath).toString();
-
-    const dir = path.join(process.cwd(), thing.d);
-    const target = path.join(dir, thing.t+'.'+thing.e);
-
-    if (fs.existsSync(target)) return console.error(target, 'already exists. Not overwriting it');
-
-    mkdirp.sync(dir);
-    const rendered = mustache.render(template, opts);
-    fs.writeFileSync(target, rendered);
-  });
+  const scaffoldFile = scaffold.bind(null, opts, __dirname)
+  things.forEach(scaffoldFile);
 
   const dirs = [
     'env',
