+++ controller/index.js
@@ -1,8 +1,5 @@
-const fs = require('fs');
-const path = require('path');
-const mkdirp = require('mkdirp');
-const mustache = require('mustache');
 const _ = require('lodash');
+const scaffold = require('../lib/scaffold');
 
 module.exports = (opts) => {
   const things = [
@@ -21,17 +18,6 @@ module.exports = (opts) => {
     {n: 'schema', d: 'schemas', t: opts.name, e: 'json'},
   ].map(thing => _.extend({p: '', d: thing.p || '', t: thing.n, e: 'js'}, thing));
 
-  things.forEach(thing => {
-    const templatePath = path.join(__dirname, thing.p, thing.n+'.mustache');
-    const template = fs.readFileSync(templatePath).toString();
-
-    const dir = path.join(process.cwd(), thing.d);
-    const target = path.join(dir, thing.t+'.'+thing.e);
-
-    if (fs.existsSync(target)) return console.error(target, 'already exists. Not overwriting it');
-
-    mkdirp.sync(dir);
-    const rendered = mustache.render(template, opts);
-    fs.writeFileSync(target, rendered);
-  });
+  const scaffoldFile = scaffold.bind(null, opts, __dirname)
+  things.forEach(scaffoldFile);
 };
